Refactor login component and add theme selector
