+++ client/src/components/ui/index.ts
@@ -26,3 +26,4 @@ export { default as SelectDropDown } from './SelectDropDown';
 export { default as MultiSelectPop } from './MultiSelectPop';
 export { default as SelectDropDownPop } from './SelectDropDownPop';
 export { default as MultiSelectDropDown } from './MultiSelectDropDown';
+export { default as ThemeSelector } from './ThemeSelector';
